# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_odatlar-bot_backend_internal_dto.CreateNotificationRequest:
    properties:
      payload:
        items:
          type: integer
        type: array
      preview:
        maxLength: 2000
        type: string
      target:
        enum:
        - all
        - selected
        type: string
      target_user_ids:
        items:
          type: integer
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
      type:
        type: string
    required:
    - title
    - type
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.NotificationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.NotificationResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.NotificationResponse:
    properties:
      createdAt:
        type: string
      deliveryCount:
        type: integer
      id:
        type: string
      payload:
        items:
          type: integer
        type: array
      preview:
        type: string
      sentAt:
        type: string
      status:
        type: string
      target:
        type: string
      targetUserIds:
        items:
          type: integer
        type: array
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UnreadCountResponse:
    properties:
      unreadCount:
        type: integer
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UpdateNotificationRequest:
    properties:
      payload:
        items:
          type: integer
        type: array
      preview:
        maxLength: 2000
        type: string
      target:
        enum:
        - all
        - selected
        type: string
      target_user_ids:
        items:
          type: integer
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
      type:
        type: string
    required:
    - target
    - title
    - type
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UserNotificationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserNotificationResponse'
        type: array
      unreadCount:
        type: integer
    type: object
  github_com_odatlar-bot_backend_internal_dto.UserNotificationResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isRead:
        type: boolean
      payload:
        items:
          type: integer
        type: array
      preview:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  github_com_odatlar-bot_backend_pkg_response.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@odatlar-bot.uz
    name: API Support
  description: Admin panel, Bot Runtime (Telegram WebApp), Qo'llanma (videolar, kurslar,
    fayllar) va Bildirishnomalar APIlari.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Odatlar Bot API
  version: "1.0"
paths:
...
  /bot-runtime/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserNotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Foydalanuvchi bildirishnomalari
      tags:
      - Notifications
  /bot-runtime/notifications/{id}/read:
    post:
      parameters:
      - description: Delivery ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UnreadCountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Bildirishnomani o'qilgan deb belgilash
      tags:
      - Notifications
  /bot-runtime/notifications/read-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UnreadCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Barcha bildirishnomalarni o'qilgan deb belgilash
      tags:
      - Notifications
...
  /bot-runtime/ws/notifications:
    get:
      description: Real-time bildirishnomalar uchun WebSocket ulanishi. initData query
        parametr orqali autentifikatsiya.
      parameters:
      - description: Telegram WebApp initData
        in: query
        name: initData
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      summary: Bildirishnomalar WebSocket
      tags:
      - Notifications
...
  /bot/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Bildirishnomalar ro'yxati (admin)
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      parameters:
      - description: Bildirishnoma ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Yangi bildirishnoma yaratish (admin)
      tags:
      - Notifications
  /bot/notifications/{id}:
    delete:
      parameters:
      - description: Bildirishnoma ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Bildirishnomani o'chirish (admin)
      tags:
      - Notifications
    get:
      parameters:
      - description: Bildirishnoma ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.NotificationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Bildirishnoma tafsilotlari (admin)
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      parameters:
      - description: Bildirishnoma ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yangilangan ma'lumotlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UpdateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Bildirishnomani yangilash (admin)
      tags:
      - Notifications
  /bot/notifications/{id}/send:
    post:
      description: Draft holatdagi bildirishnomani foydalanuvchilarga yuboradi
      parameters:
      - description: Bildirishnoma ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Bildirishnomani yuborish (admin)
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  BotToken:
    description: 'Telegram bot token. Format: Bearer {bot_token}'
    in: header
    name: Authorization
    type: apiKey
  TelegramInitData:
    description: 'Telegram WebApp initData. Alternativa: Authorization: tma {initData}'
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"